refactor(intersection): divide source files and modifyPathVelocity (#6134)
