+++ planning/behavior_velocity_intersection_module/CMakeLists.txt
@@ -8,11 +8,16 @@ pluginlib_export_plugin_description_file(behavior_velocity_planner plugins.xml)
 find_package(OpenCV REQUIRED)
 
 ament_auto_add_library(${PROJECT_NAME} SHARED
-  src/debug.cpp
   src/manager.cpp
-  src/scene_intersection.cpp
-  src/scene_merge_from_private_road.cpp
   src/util.cpp
+  src/scene_intersection.cpp
+  src/intersection_lanelets.cpp
+  src/scene_intersection_prepare_data.cpp
+  src/scene_intersection_stuck.cpp
+  src/scene_intersection_occlusion.cpp
+  src/scene_intersection_collision.cpp
+  src/scene_merge_from_private_road.cpp
+  src/debug.cpp
 )
 
 target_link_libraries(${PROJECT_NAME}
