+++ planning/behavior_velocity_intersection_module/src/manager.hpp
@@ -44,7 +44,6 @@ private:
   IntersectionModule::PlannerParam intersection_param_;
   // additional for INTERSECTION_OCCLUSION
   RTCInterface occlusion_rtc_interface_;
-  std::unordered_map<int64_t, UUID> occlusion_map_uuid_;
 
   void launchNewModules(const autoware_auto_planning_msgs::msg::PathWithLaneId & path) override;
