+++ planning/behavior_velocity_intersection_module/src/scene_intersection.hpp
@@ -15,7 +15,11 @@
 #ifndef SCENE_INTERSECTION_HPP_
 #define SCENE_INTERSECTION_HPP_
 
-#include "util_type.hpp"
+#include "decision_result.hpp"
+#include "interpolated_path_info.hpp"
+#include "intersection_lanelets.hpp"
+#include "intersection_stoplines.hpp"
+#include "result.hpp"
 
 #include <behavior_velocity_planner_common/scene_module_interface.hpp>
 #include <behavior_velocity_planner_common/utilization/state_machine.hpp>
@@ -26,270 +30,20 @@
 #include <std_msgs/msg/string.hpp>
 #include <tier4_debug_msgs/msg/float64_multi_array_stamped.hpp>
 
-#include <lanelet2_core/LaneletMap.h>
-#include <lanelet2_routing/RoutingGraph.h>
+#include <lanelet2_core/Forward.h>
+#include <lanelet2_core/primitives/LineString.h>
+#include <lanelet2_routing/Forward.h>
 
-#include <algorithm>
-#include <map>
 #include <memory>
+#include <optional>
 #include <set>
 #include <string>
+#include <tuple>
 #include <utility>
-#include <variant>
 #include <vector>
 
 namespace behavior_velocity_planner
 {
-
-using TimeDistanceArray = std::vector<std::pair<double /* time*/, double /* distance*/>>;
-
-struct DebugData
-{
-  std::optional<geometry_msgs::msg::Pose> collision_stop_wall_pose{std::nullopt};
-  std::optional<geometry_msgs::msg::Pose> occlusion_stop_wall_pose{std::nullopt};
-  std::optional<geometry_msgs::msg::Pose> occlusion_first_stop_wall_pose{std::nullopt};
-  std::optional<geometry_msgs::msg::Pose> first_pass_judge_wall_pose{std::nullopt};
-  bool passed_first_pass_judge{false};
-  bool passed_second_pass_judge{false};
-  std::optional<geometry_msgs::msg::Pose> second_pass_judge_wall_pose{std::nullopt};
-  std::optional<std::vector<lanelet::CompoundPolygon3d>> attention_area{std::nullopt};
-  std::optional<std::vector<lanelet::CompoundPolygon3d>> occlusion_attention_area{std::nullopt};
-  std::optional<lanelet::CompoundPolygon3d> ego_lane{std::nullopt};
-  std::optional<std::vector<lanelet::CompoundPolygon3d>> adjacent_area{std::nullopt};
-  std::optional<lanelet::CompoundPolygon3d> first_attention_area{std::nullopt};
-  std::optional<lanelet::CompoundPolygon3d> second_attention_area{std::nullopt};
-  std::optional<geometry_msgs::msg::Polygon> stuck_vehicle_detect_area{std::nullopt};
-  std::optional<std::vector<lanelet::CompoundPolygon3d>> yield_stuck_detect_area{std::nullopt};
-  std::optional<geometry_msgs::msg::Polygon> candidate_collision_ego_lane_polygon{std::nullopt};
-  std::vector<geometry_msgs::msg::Polygon> candidate_collision_object_polygons;
-  autoware_auto_perception_msgs::msg::PredictedObjects conflicting_targets;
-  autoware_auto_perception_msgs::msg::PredictedObjects amber_ignore_targets;
-  autoware_auto_perception_msgs::msg::PredictedObjects red_overshoot_ignore_targets;
-  autoware_auto_perception_msgs::msg::PredictedObjects stuck_targets;
-  autoware_auto_perception_msgs::msg::PredictedObjects yield_stuck_targets;
-  std::vector<geometry_msgs::msg::Polygon> occlusion_polygons;
-  std::optional<std::pair<geometry_msgs::msg::Point, geometry_msgs::msg::Point>>
-    nearest_occlusion_projection{std::nullopt};
-  autoware_auto_perception_msgs::msg::PredictedObjects blocking_attention_objects;
-  std::optional<geometry_msgs::msg::Pose> absence_traffic_light_creep_wall{std::nullopt};
-  std::optional<double> static_occlusion_with_traffic_light_timeout{std::nullopt};
-};
-
-/**
- * @struct
- * @brief see the document for more details of IntersectionLanelets
- */
-struct IntersectionLanelets
-{
-public:
-  /**
-   * update conflicting lanelets and traffic priority information
-   */
-  void update(
-    const bool is_prioritized, const util::InterpolatedPathInfo & interpolated_path_info,
-    const tier4_autoware_utils::LinearRing2d & footprint, const double vehicle_length,
-    lanelet::routing::RoutingGraphPtr routing_graph_ptr);
-
-  const lanelet::ConstLanelets & attention() const
-  {
-    return is_prioritized_ ? attention_non_preceding_ : attention_;
-  }
-  const std::vector<std::optional<lanelet::ConstLineString3d>> & attention_stoplines() const
-  {
-    return is_prioritized_ ? attention_non_preceding_stoplines_ : attention_stoplines_;
-  }
-  const lanelet::ConstLanelets & conflicting() const { return conflicting_; }
-  const lanelet::ConstLanelets & adjacent() const { return adjacent_; }
-  const lanelet::ConstLanelets & occlusion_attention() const
-  {
-    return is_prioritized_ ? attention_non_preceding_ : occlusion_attention_;
-  }
-  const lanelet::ConstLanelets & attention_non_preceding() const
-  {
-    return attention_non_preceding_;
-  }
-  const std::vector<lanelet::CompoundPolygon3d> & attention_area() const
-  {
-    return is_prioritized_ ? attention_non_preceding_area_ : attention_area_;
-  }
-  const std::vector<lanelet::CompoundPolygon3d> & conflicting_area() const
-  {
-    return conflicting_area_;
-  }
-  const std::vector<lanelet::CompoundPolygon3d> & adjacent_area() const { return adjacent_area_; }
-  const std::vector<lanelet::CompoundPolygon3d> & occlusion_attention_area() const
-  {
-    return occlusion_attention_area_;
-  }
-  const std::optional<lanelet::ConstLanelet> & first_conflicting_lane() const
-  {
-    return first_conflicting_lane_;
-  }
-  const std::optional<lanelet::CompoundPolygon3d> & first_conflicting_area() const
-  {
-    return first_conflicting_area_;
-  }
-  const std::optional<lanelet::ConstLanelet> & first_attention_lane() const
-  {
-    return first_attention_lane_;
-  }
-  const std::optional<lanelet::CompoundPolygon3d> & first_attention_area() const
-  {
-    return first_attention_area_;
-  }
-  const std::optional<lanelet::ConstLanelet> & second_attention_lane() const
-  {
-    return second_attention_lane_;
-  }
-  const std::optional<lanelet::CompoundPolygon3d> & second_attention_area() const
-  {
-    return second_attention_area_;
-  }
-
-  /**
-   * the set of attention lanelets which is topologically merged
-   */
-  lanelet::ConstLanelets attention_;
-  std::vector<lanelet::CompoundPolygon3d> attention_area_;
-
-  /**
-   * the stop lines for each attention_ lanelets
-   */
-  std::vector<std::optional<lanelet::ConstLineString3d>> attention_stoplines_;
-
-  /**
-   * the conflicting part of attention lanelets
-   */
-  lanelet::ConstLanelets attention_non_preceding_;
-  std::vector<lanelet::CompoundPolygon3d> attention_non_preceding_area_;
-
-  /**
-   * the stop lines for each attention_non_preceding_
-   */
-  std::vector<std::optional<lanelet::ConstLineString3d>> attention_non_preceding_stoplines_;
-
-  /**
-   * the conflicting lanelets of the objective intersection lanelet
-   */
-  lanelet::ConstLanelets conflicting_;
-  std::vector<lanelet::CompoundPolygon3d> conflicting_area_;
-
-  /**
-   *
-   */
-  lanelet::ConstLanelets adjacent_;
-  std::vector<lanelet::CompoundPolygon3d> adjacent_area_;
-
-  /**
-   * the set of attention lanelets for occlusion detection which is topologically merged
-   */
-  lanelet::ConstLanelets occlusion_attention_;
-  std::vector<lanelet::CompoundPolygon3d> occlusion_attention_area_;
-
-  /**
-   * the vector of sum of each occlusion_attention lanelet
-   */
-  std::vector<double> occlusion_attention_size_;
-
-  /**
-   * the first conflicting lanelet which ego path points intersect for the first time
-   */
-  std::optional<lanelet::ConstLanelet> first_conflicting_lane_{std::nullopt};
-  std::optional<lanelet::CompoundPolygon3d> first_conflicting_area_{std::nullopt};
-
-  /**
-   * the first attention lanelet which ego path points intersect for the first time
-   */
-  std::optional<lanelet::ConstLanelet> first_attention_lane_{std::nullopt};
-  std::optional<lanelet::CompoundPolygon3d> first_attention_area_{std::nullopt};
-
-  /**
-   * the second attention lanelet which ego path points intersect next to the
-   * first_attention_lanelet
-   */
-  bool second_attention_lane_empty_{false};
-  std::optional<lanelet::ConstLanelet> second_attention_lane_{std::nullopt};
-  std::optional<lanelet::CompoundPolygon3d> second_attention_area_{std::nullopt};
-
-  /**
-   * flag if the intersection is prioritized by the traffic light
-   */
-  bool is_prioritized_{false};
-};
-
-/**
- * @struct
- * @brief see the document for more details of IntersectionStopLines
- */
-struct IntersectionStopLines
-{
-  size_t closest_idx{0};
-
-  /**
-   * stuck_stopline is null if ego path does not intersect with first_conflicting_area
-   */
-  std::optional<size_t> stuck_stopline{std::nullopt};
-
-  /**
-   * default_stopline is null if it is calculated negative from first_attention_stopline
-   */
-  std::optional<size_t> default_stopline{std::nullopt};
-
-  /**
-   * first_attention_stopline is null if ego footprint along the path does not intersect with
-   * attention area. if path[0] satisfies the condition, it is 0
-   */
-  std::optional<size_t> first_attention_stopline{std::nullopt};
-
-  /**
-   * second_attention_stopline is null if ego footprint along the path does not intersect with
-   * second_attention_lane. if path[0] satisfies the condition, it is 0
-   */
-  std::optional<size_t> second_attention_stopline{std::nullopt};
-
-  /**
-   * occlusion_peeking_stopline is null if path[0] is already inside the attention area
-   */
-  std::optional<size_t> occlusion_peeking_stopline{std::nullopt};
-
-  /**
-   * first_pass_judge_line is before first_attention_stopline by the braking distance. if its value
-   * is calculated negative, it is 0
-   */
-  size_t first_pass_judge_line{0};
-
-  /**
-   * second_pass_judge_line is before second_attention_stopline by the braking distance. if
-   * second_attention_lane is null, it is same as first_pass_judge_line
-   */
-  size_t second_pass_judge_line{0};
-
-  /**
-   * occlusion_wo_tl_pass_judge_line is null if ego footprint along the path does not intersect with
-   * the centerline of the first_attention_lane
-   */
-  size_t occlusion_wo_tl_pass_judge_line{0};
-};
-
-/**
- * @struct
- * @brief see the document for more details of PathLanelets
- */
-struct PathLanelets
-{
-  lanelet::ConstLanelets prev;
-  // lanelet::ConstLanelet entry2ego; this is included in `all` if exists
-  lanelet::ConstLanelet
-    ego_or_entry2exit;  // this is `assigned lane` part of the path(not from
-                        // ego) if ego is before the intersection, otherwise from ego to exit
-  std::optional<lanelet::ConstLanelet> next =
-    std::nullopt;  // this is nullopt is the goal is inside intersection
-  lanelet::ConstLanelets all;
-  lanelet::ConstLanelets
-    conflicting_interval_and_remaining;  // the left/right-most interval of path conflicting with
-                                         // conflicting lanelets plus the next lane part of the
-                                         // path
-};
 
 /**
  * @struct
@@ -315,19 +69,6 @@ struct TargetObjects
   std::vector<TargetObject> parked_attention_objects;
   std::vector<TargetObject> intersection_area_objects;
   std::vector<TargetObject> all_attention_objects;  // TODO(Mamoru Sobue): avoid copy
-};
-
-/**
- * @struct
- * @brief categorize traffic light priority
- */
-enum class TrafficPrioritizedLevel {
-  //! The target lane's traffic signal is red or the ego's traffic signal has an arrow.
-  FULLY_PRIORITIZED = 0,
-  //! The target lane's traffic signal is amber
-  PARTIALLY_PRIORITIZED,
-  //! The target lane's traffic signal is green
-  NOT_PRIORITIZED
 };
 
 class IntersectionModule : public SceneModuleInterface
@@ -457,144 +198,77 @@ public:
     RTC_OCCLUDED,
   };
 
+  struct DebugData
+  {
+    std::optional<geometry_msgs::msg::Pose> collision_stop_wall_pose{std::nullopt};
+    std::optional<geometry_msgs::msg::Pose> occlusion_stop_wall_pose{std::nullopt};
+    std::optional<geometry_msgs::msg::Pose> occlusion_first_stop_wall_pose{std::nullopt};
+    std::optional<geometry_msgs::msg::Pose> first_pass_judge_wall_pose{std::nullopt};
+    bool passed_first_pass_judge{false};
+    bool passed_second_pass_judge{false};
+    std::optional<geometry_msgs::msg::Pose> second_pass_judge_wall_pose{std::nullopt};
+    std::optional<std::vector<lanelet::CompoundPolygon3d>> attention_area{std::nullopt};
+    std::optional<std::vector<lanelet::CompoundPolygon3d>> occlusion_attention_area{std::nullopt};
+    std::optional<lanelet::CompoundPolygon3d> ego_lane{std::nullopt};
+    std::optional<std::vector<lanelet::CompoundPolygon3d>> adjacent_area{std::nullopt};
+    std::optional<lanelet::CompoundPolygon3d> first_attention_area{std::nullopt};
+    std::optional<lanelet::CompoundPolygon3d> second_attention_area{std::nullopt};
+    std::optional<geometry_msgs::msg::Polygon> stuck_vehicle_detect_area{std::nullopt};
+    std::optional<std::vector<lanelet::CompoundPolygon3d>> yield_stuck_detect_area{std::nullopt};
+    std::optional<geometry_msgs::msg::Polygon> candidate_collision_ego_lane_polygon{std::nullopt};
+    std::vector<geometry_msgs::msg::Polygon> candidate_collision_object_polygons;
+    autoware_auto_perception_msgs::msg::PredictedObjects conflicting_targets;
+    autoware_auto_perception_msgs::msg::PredictedObjects amber_ignore_targets;
+    autoware_auto_perception_msgs::msg::PredictedObjects red_overshoot_ignore_targets;
+    autoware_auto_perception_msgs::msg::PredictedObjects stuck_targets;
+    autoware_auto_perception_msgs::msg::PredictedObjects yield_stuck_targets;
+    std::vector<geometry_msgs::msg::Polygon> occlusion_polygons;
+    std::optional<std::pair<geometry_msgs::msg::Point, geometry_msgs::msg::Point>>
+      nearest_occlusion_projection{std::nullopt};
+    autoware_auto_perception_msgs::msg::PredictedObjects blocking_attention_objects;
+    std::optional<geometry_msgs::msg::Pose> absence_traffic_light_creep_wall{std::nullopt};
+    std::optional<double> static_occlusion_with_traffic_light_timeout{std::nullopt};
+  };
+
+  using TimeDistanceArray = std::vector<std::pair<double /* time*/, double /* distance*/>>;
+
   /**
    * @struct
-   * @brief Internal error or ego already passed pass_judge_line
+   * @brief categorize traffic light priority
    */
-  struct Indecisive
-  {
-    std::string error;
+  enum class TrafficPrioritizedLevel {
+    //! The target lane's traffic signal is red or the ego's traffic signal has an arrow.
+    FULLY_PRIORITIZED = 0,
+    //! The target lane's traffic signal is amber
+    PARTIALLY_PRIORITIZED,
+    //! The target lane's traffic signal is green
+    NOT_PRIORITIZED
   };
-  /**
-   * @struct
-   * @brief detected stuck vehicle
-   */
-  struct StuckStop
-  {
-    size_t closest_idx{0};
-    size_t stuck_stopline_idx{0};
-    std::optional<size_t> occlusion_stopline_idx{std::nullopt};
-  };
-  /**
-   * @struct
-   * @brief yielded by vehicle on the attention area
-   */
-  struct YieldStuckStop
-  {
-    size_t closest_idx{0};
-    size_t stuck_stopline_idx{0};
-  };
-  /**
-   * @struct
-   * @brief only collision is detected
-   */
-  struct NonOccludedCollisionStop
-  {
-    size_t closest_idx{0};
-    size_t collision_stopline_idx{0};
-    size_t occlusion_stopline_idx{0};
-  };
-  /**
-   * @struct
-   * @brief occlusion is detected so ego needs to stop at the default stop line position
-   */
-  struct FirstWaitBeforeOcclusion
-  {
-    bool is_actually_occlusion_cleared{false};
-    size_t closest_idx{0};
-    size_t first_stopline_idx{0};
-    size_t occlusion_stopline_idx{0};
-  };
-  /**
-   * @struct
-   * @brief ego is approaching the boundary of attention area in the presence of traffic light
-   */
-  struct PeekingTowardOcclusion
-  {
-    //! if intersection_occlusion is disapproved externally through RTC, it indicates
-    //! "is_forcefully_occluded"
-    bool is_actually_occlusion_cleared{false};
-    bool temporal_stop_before_attention_required{false};
-    size_t closest_idx{0};
-    size_t collision_stopline_idx{0};
-    size_t first_attention_stopline_idx{0};
-    size_t occlusion_stopline_idx{0};
-    //! if null, it is dynamic occlusion and shows up intersection_occlusion(dyn). if valid, it
-    //! contains the remaining time to release the static occlusion stuck and shows up
-    //! intersection_occlusion(x.y)
-    std::optional<double> static_occlusion_timeout{std::nullopt};
-  };
-  /**
-   * @struct
-   * @brief both collision and occlusion are detected in the presence of traffic light
-   */
-  struct OccludedCollisionStop
-  {
-    bool is_actually_occlusion_cleared{false};
-    bool temporal_stop_before_attention_required{false};
-    size_t closest_idx{0};
-    size_t collision_stopline_idx{0};
-    size_t first_attention_stopline_idx{0};
-    size_t occlusion_stopline_idx{0};
-    //! if null, it is dynamic occlusion and shows up intersection_occlusion(dyn). if valid, it
-    //! contains the remaining time to release the static occlusion stuck
-    std::optional<double> static_occlusion_timeout{std::nullopt};
-  };
-  /**
-   * @struct
-   * @brief at least occlusion is detected in the absence of traffic light
-   */
-  struct OccludedAbsenceTrafficLight
-  {
-    bool is_actually_occlusion_cleared{false};
-    bool collision_detected{false};
-    bool temporal_stop_before_attention_required{false};
-    size_t closest_idx{0};
-    size_t first_attention_area_stopline_idx{0};
-    size_t peeking_limit_line_idx{0};
-  };
-  /**
-   * @struct
-   * @brief both collision and occlusion are not detected
-   */
-  struct Safe
-  {
-    size_t closest_idx{0};
-    size_t collision_stopline_idx{0};
-    size_t occlusion_stopline_idx{0};
-  };
-  /**
-   * @struct
-   * @brief traffic light is red or arrow signal
-   */
-  struct FullyPrioritized
-  {
-    bool collision_detected{false};
-    size_t closest_idx{0};
-    size_t collision_stopline_idx{0};
-    size_t occlusion_stopline_idx{0};
-  };
-  using DecisionResult = std::variant<
-    Indecisive,                   //! internal process error, or over the pass judge line
-    StuckStop,                    //! detected stuck vehicle
-    YieldStuckStop,               //! detected yield stuck vehicle
-    NonOccludedCollisionStop,     //! detected collision while FOV is clear
-    FirstWaitBeforeOcclusion,     //! stop for a while before peeking to occlusion
-    PeekingTowardOcclusion,       //! peeking into occlusion while collision is not detected
-    OccludedCollisionStop,        //! occlusion and collision are both detected
-    OccludedAbsenceTrafficLight,  //! occlusion is detected in the absence of traffic light
-    Safe,                         //! judge as safe
-    FullyPrioritized              //! only detect vehicles violating traffic rules
-    >;
+  /** @} */
 
   IntersectionModule(
     const int64_t module_id, const int64_t lane_id, std::shared_ptr<const PlannerData> planner_data,
     const PlannerParam & planner_param, const std::set<lanelet::Id> & associative_ids,
-    const std::string & turn_direction, const bool has_traffic_light,
-    const bool enable_occlusion_detection, rclcpp::Node & node, const rclcpp::Logger logger,
-    const rclcpp::Clock::SharedPtr clock);
+    const std::string & turn_direction, const bool has_traffic_light, rclcpp::Node & node,
+    const rclcpp::Logger logger, const rclcpp::Clock::SharedPtr clock);
 
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup primary-function [fn] primary functions
+   * the entrypoint of this module is modifyPathVelocity() function that calculates safety decision
+   * of latest context and send it to RTC and then react to RTC approval. The reaction to RTC
+   * approval may not be based on the latest decision of this module depending on the auto-mode
+   * configuration. For module side it is not visible if the module is operating in auto-mode or
+   * manual-module. At first, initializeRTCStatus() is called to reset the safety value of
+   * INTERSECTION and INTERSECTION_OCCLUSION. Then modifyPathVelocityDetail() is called to analyze
+   * the context. Then prepareRTCStatus() is called to set the safety value of INTERSECTION and
+   * INTERSECTION_OCCLUSION.
+   * @{
+   */
   bool modifyPathVelocity(PathWithLaneId * path, StopReason * stop_reason) override;
+  /** @}*/
 
   visualization_msgs::msg::MarkerArray createDebugMarkerArray() override;
   motion_utils::VirtualWalls createVirtualWalls() override;
@@ -609,137 +283,421 @@ public:
 
 private:
   rclcpp::Node & node_;
+
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup const-variables [var] const variables
+   * following variables are unique to this intersection lanelet or to this module
+   * @{
+   */
+  //! lanelet of this intersection
   const lanelet::Id lane_id_;
+
+  //! associative(sibling) lanelets ids
   const std::set<lanelet::Id> associative_ids_;
+
+  //! turn_direction of this lane
   const std::string turn_direction_;
+
+  //! flag if this intersection is traffic controlled
   const bool has_traffic_light_;
 
-  // Parameter
+  //! RTC uuid for INTERSECTION_OCCLUSION
+  const UUID occlusion_uuid_;
+  /** @}*/
+
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup semi-const-variables [var] semi-const variables
+   * following variables are immutable once initialized
+   * @{
+   */
   PlannerParam planner_param_;
 
-  std::optional<IntersectionLanelets> intersection_lanelets_{std::nullopt};
+  //! cache IntersectionLanelets struct
+  std::optional<intersection::IntersectionLanelets> intersection_lanelets_{std::nullopt};
 
-  // for pass judge decision
-  bool is_go_out_{false};
+  //! cache discretized occlusion detection lanelets
+  std::optional<std::vector<lanelet::ConstLineString3d>> occlusion_attention_divisions_{
+    std::nullopt};
+  /** @}*/
+
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup pass-judge-variable [var] pass judge variables
+   * following variables are state variables that depends on how the vehicle passed the intersection
+   * @{
+   */
+  //! if true, this module never commands to STOP anymore
   bool is_permanent_go_{false};
-  DecisionResult prev_decision_result_{Indecisive{""}};
-  OcclusionType prev_occlusion_status_;
+
+  //! for checking if ego is over the pass judge lines because previously the situation was SAFE
+  intersection::DecisionResult prev_decision_result_{intersection::Indecisive{""}};
+
+  //! flag if ego passed the 1st_pass_judge_line while peeking. If this is true, 1st_pass_judge_line
+  //! is treated as the same position as occlusion_peeking_stopline
   bool passed_1st_judge_line_while_peeking_{false};
+
+  //! save the time when ego passed the 1st/2nd_pass_judge_line with safe decision. If collision is
+  //! expected after these variables are non-null, then it is the fault of past perception failure
+  //! at these time.
   std::optional<rclcpp::Time> safely_passed_1st_judge_line_time_{std::nullopt};
   std::optional<rclcpp::Time> safely_passed_2nd_judge_line_time_{std::nullopt};
+  /** @}*/
 
-  // for occlusion detection
-  const bool enable_occlusion_detection_;
-  std::optional<std::vector<lanelet::ConstLineString3d>> occlusion_attention_divisions_{
-    std::nullopt};                        //! for caching discretized occlusion detection lanelets
-  StateMachine collision_state_machine_;  //! for stable collision checking
-  StateMachine before_creep_state_machine_;    //! for two phase stop
-  StateMachine occlusion_stop_state_machine_;  //! for stable occlusion detection
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup collision-variables [var] collision detection
+   * @{
+   */
+  //! debouncing for stable SAFE decision
+  StateMachine collision_state_machine_;
+  /** @} */
+
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup occlusion-variables [var] occlusion detection variables
+   * @{
+   */
+  OcclusionType prev_occlusion_status_;
+
+  //! debouncing for the first brief stop at the default stopline
+  StateMachine before_creep_state_machine_;
+
+  //! debouncing for stable CLEARED decision
+  StateMachine occlusion_stop_state_machine_;
+
+  //! debouncing for the brief stop at the boundary of attention area(if required by the flag)
   StateMachine temporal_stop_before_attention_state_machine_;
+
+  //! time counter for the stuck detection due to occlusion caused static objects
   StateMachine static_occlusion_timeout_state_machine_;
+  /** @} */
 
   std::optional<rclcpp::Time> initial_green_light_observed_time_{std::nullopt};
 
-  // for RTC
-  const UUID occlusion_uuid_;
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup RTC-variables [var] RTC variables
+   *
+   * intersection module has additional rtc_interface_ for INTERSECTION_OCCLUSION in addition to the
+   * default rtc_interface of SceneModuleManagerInterfaceWithRTC. activated_ is the derived member
+   * of this module which is updated by the RTC config/service, so it should be read-only in this
+   * module. occlusion_safety_ and occlusion_stop_distance_ are the corresponding RTC value for
+   * INTERSECTION_OCCLUSION.
+   * @{
+   */
   bool occlusion_safety_{true};
   double occlusion_stop_distance_{0.0};
   bool occlusion_activated_{true};
   bool occlusion_first_stop_required_{false};
+  /** @}*/
 
+private:
   /**
-   * @fn
-   * @brief set all RTC variable to safe and -inf
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @ingroup primary-functions
+   * @{
+   */
+  /**
+   * @brief set all RTC variable to true(safe) and -INF
    */
   void initializeRTCStatus();
+
   /**
-   * @fn
    * @brief analyze traffic_light/occupancy/objects context and return DecisionResult
    */
-  DecisionResult modifyPathVelocityDetail(PathWithLaneId * path, StopReason * stop_reason);
+  intersection::DecisionResult modifyPathVelocityDetail(
+    PathWithLaneId * path, StopReason * stop_reason);
+
   /**
-   * @fn
    * @brief set RTC value according to calculated DecisionResult
    */
   void prepareRTCStatus(
-    const DecisionResult &, const autoware_auto_planning_msgs::msg::PathWithLaneId & path);
+    const intersection::DecisionResult &,
+    const autoware_auto_planning_msgs::msg::PathWithLaneId & path);
 
   /**
-   * @fn
-   * @brief find TrafficPrioritizedLevel
+   * @brief act based on current RTC approval
    */
-  TrafficPrioritizedLevel getTrafficPrioritizedLevel(lanelet::ConstLanelet lane);
+  void reactRTCApproval(
+    const intersection::DecisionResult & decision_result,
+    autoware_auto_planning_msgs::msg::PathWithLaneId * path, StopReason * stop_reason);
+  /** @}*/
 
+private:
   /**
-   * @fn
-   * @brief generate IntersectionLanelets
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup prepare-data [fn] basic data construction
+   * @{
    */
-  IntersectionLanelets getObjectiveLanelets(
-    lanelet::LaneletMapConstPtr lanelet_map_ptr,
-    lanelet::routing::RoutingGraphPtr routing_graph_ptr,
-    const lanelet::ConstLanelet assigned_lanelet, const lanelet::ConstLanelets & lanelets_on_path);
+  /**
+   * @struct
+   */
+  struct BasicData
+  {
+    intersection::InterpolatedPathInfo interpolated_path_info;
+    intersection::IntersectionStopLines intersection_stoplines;
+    intersection::PathLanelets path_lanelets;
+  };
 
   /**
-   * @fn
+   * @brief prepare basic data structure
+   * @return return IntersectionStopLines if all data is valid, otherwise Indecisive
+   * @note if successful, it is ensure that intersection_lanelets_,
+   * intersection_lanelets.first_conflicting_lane are not null
+   *
+   * To simplify modifyPathVelocityDetail(), this function is used at first
+   */
+  intersection::Result<BasicData, intersection::Indecisive> prepareIntersectionData(
+    const bool is_prioritized, PathWithLaneId * path);
+
+  /**
+   * @brief find the associated stopline road marking of assigned lanelet
+   */
+  std::optional<size_t> getStopLineIndexFromMap(
+    const intersection::InterpolatedPathInfo & interpolated_path_info,
+    lanelet::ConstLanelet assigned_lanelet);
+
+  /**
    * @brief generate IntersectionStopLines
    */
-  std::optional<IntersectionStopLines> generateIntersectionStopLines(
+  std::optional<intersection::IntersectionStopLines> generateIntersectionStopLines(
     lanelet::ConstLanelet assigned_lanelet,
     const lanelet::CompoundPolygon3d & first_conflicting_area,
     const lanelet::ConstLanelet & first_attention_lane,
     const std::optional<lanelet::CompoundPolygon3d> & second_attention_area_opt,
-    const util::InterpolatedPathInfo & interpolated_path_info,
+    const intersection::InterpolatedPathInfo & interpolated_path_info,
     autoware_auto_planning_msgs::msg::PathWithLaneId * original_path);
 
   /**
-   * @fn
-   * @brief find the associated stopline road marking of assigned lanelet
+   * @brief generate IntersectionLanelets
    */
-  std::optional<size_t> getStopLineIndexFromMap(
-    const util::InterpolatedPathInfo & interpolated_path_info,
-    lanelet::ConstLanelet assigned_lanelet);
+  intersection::IntersectionLanelets generateObjectiveLanelets(
+    lanelet::LaneletMapConstPtr lanelet_map_ptr,
+    lanelet::routing::RoutingGraphPtr routing_graph_ptr,
+    const lanelet::ConstLanelet assigned_lanelet);
 
   /**
-   * @fn
    * @brief generate PathLanelets
    */
-  std::optional<PathLanelets> generatePathLanelets(
+  std::optional<intersection::PathLanelets> generatePathLanelets(
     const lanelet::ConstLanelets & lanelets_on_path,
-    const util::InterpolatedPathInfo & interpolated_path_info,
+    const intersection::InterpolatedPathInfo & interpolated_path_info,
     const lanelet::CompoundPolygon3d & first_conflicting_area,
     const std::vector<lanelet::CompoundPolygon3d> & conflicting_areas,
     const std::optional<lanelet::CompoundPolygon3d> & first_attention_area,
     const std::vector<lanelet::CompoundPolygon3d> & attention_areas, const size_t closest_idx);
 
   /**
-   * @fn
-   * @brief check stuck
+   * @brief generate discretized detection lane linestring.
    */
-  bool checkStuckVehicleInIntersection(const PathLanelets & path_lanelets, DebugData * debug_data);
+  std::vector<lanelet::ConstLineString3d> generateDetectionLaneDivisions(
+    lanelet::ConstLanelets detection_lanelets,
+    const lanelet::routing::RoutingGraphPtr routing_graph_ptr, const double resolution);
+  /** @} */
+
+private:
+  /**
+   * @defgroup utility [fn] utility member function
+   * @{
+   */
+  void stoppedAtPositionForDuration(
+    const autoware_auto_planning_msgs::msg::PathWithLaneId & path, const size_t position,
+    const double duration, StateMachine * state_machine);
+  /** @} */
+
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup get-traffic-light [fn] traffic light
+   * @{
+   */
+  /**
+   * @brief check if associated traffic light is green
+   */
+  bool isGreenSolidOn() const;
 
   /**
-   * @fn
-   * @brief check yield stuck
+   * @brief find TrafficPrioritizedLevel
    */
-  bool checkYieldStuckVehicleInIntersection(
-    const TargetObjects & target_objects, const util::InterpolatedPathInfo & interpolated_path_info,
-    const lanelet::ConstLanelets & attention_lanelets, DebugData * debug_data);
+  TrafficPrioritizedLevel getTrafficPrioritizedLevel() const;
+  /** @} */
 
+private:
   /**
-   * @fn
    * @brief categorize target objects
    */
   TargetObjects generateTargetObjects(
-    const IntersectionLanelets & intersection_lanelets,
+    const intersection::IntersectionLanelets & intersection_lanelets,
     const std::optional<Polygon2d> & intersection_area) const;
 
+private:
   /**
-   * @fn
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup yield [fn] check stuck
+   * @{
+   */
+  /**
+   * @brief check stuck status
+   * @attention this function has access to value() of intersection_lanelets_,
+   * intersection_lanelets.first_conflicting_lane(). They are ensured in prepareIntersectionData()
+   */
+  std::optional<intersection::StuckStop> isStuckStatus(
+    const autoware_auto_planning_msgs::msg::PathWithLaneId & path,
+    const intersection::IntersectionStopLines & intersection_stoplines,
+    const intersection::PathLanelets & path_lanelets) const;
+
+  bool isTargetStuckVehicleType(
+    const autoware_auto_perception_msgs::msg::PredictedObject & object) const;
+
+  /**
+   * @brief check stuck
+   */
+  bool checkStuckVehicleInIntersection(const intersection::PathLanelets & path_lanelets) const;
+  /** @} */
+
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup yield [fn] check yield stuck
+   * @{
+   */
+  /**
+   * @brief check yield stuck status
+   * @attention this function has access to value() of intersection_lanelets_,
+   * intersection_stoplines.default_stopline, intersection_stoplines.first_attention_stopline
+   */
+  std::optional<intersection::YieldStuckStop> isYieldStuckStatus(
+    const autoware_auto_planning_msgs::msg::PathWithLaneId & path,
+    const intersection::InterpolatedPathInfo & interpolated_path_info,
+    const intersection::IntersectionStopLines & intersection_stoplines,
+    const TargetObjects & target_objects) const;
+
+  /**
+   * @brief check yield stuck
+   */
+  bool checkYieldStuckVehicleInIntersection(
+    const TargetObjects & target_objects,
+    const intersection::InterpolatedPathInfo & interpolated_path_info,
+    const lanelet::ConstLanelets & attention_lanelets) const;
+  /** @} */
+
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup occlusion [fn] check occlusion
+   * @{
+   */
+  /**
+   * @brief check occlusion status
+   * @attention this function has access to value() of occlusion_attention_divisions_,
+   * intersection_lanelets.first_attention_area()
+   */
+  std::tuple<
+    OcclusionType, bool /* module detection with margin */,
+    bool /* reconciled occlusion disapproval */>
+  getOcclusionStatus(
+    const TrafficPrioritizedLevel & traffic_prioritized_level,
+    const intersection::InterpolatedPathInfo & interpolated_path_info,
+    const intersection::IntersectionLanelets & intersection_lanelets,
+    const TargetObjects & target_objects);
+
+  /**
+   * @brief calculate detected occlusion status(NOT | STATICALLY | DYNAMICALLY)
+   */
+  OcclusionType detectOcclusion(
+    const std::vector<lanelet::CompoundPolygon3d> & attention_areas,
+    const lanelet::ConstLanelets & adjacent_lanelets,
+    const lanelet::CompoundPolygon3d & first_attention_area,
+    const intersection::InterpolatedPathInfo & interpolated_path_info,
+    const std::vector<lanelet::ConstLineString3d> & lane_divisions,
+    const TargetObjects & target_objects);
+  /** @} */
+
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup pass-judge-decision [fn] pass judge decision
+   * @{
+   */
+  /**
+   * @brief check if ego is already over the pass judge line
+   * @return if ego is over both 1st/2nd pass judge lines, return Indecisive, else return
+   * (is_over_1st_pass_judge, is_over_2nd_pass_judge)
+   * @attention this function has access to value() of intersection_stoplines.default_stopline,
+   * intersection_stoplines.occlusion_stopline
+   */
+  intersection::Result<
+    intersection::Indecisive,
+    std::pair<bool /* is_over_1st_pass_judge */, bool /* is_over_2nd_pass_judge */>>
+  isOverPassJudgeLinesStatus(
+    const autoware_auto_planning_msgs::msg::PathWithLaneId & path, const bool is_occlusion_state,
+    const intersection::IntersectionStopLines & intersection_stoplines);
+  /** @} */
+
+private:
+  /**
+   ***********************************************************
+   ***********************************************************
+   ***********************************************************
+   * @defgroup collision-detection [fn] check collision
+   * @{
+   */
+  bool isTargetCollisionVehicleType(
+    const autoware_auto_perception_msgs::msg::PredictedObject & object) const;
+
+  /**
+   * @brief check if there are any objects around the stoplines on the attention areas when ego
+   * entered the intersection on green light
+   * @return return NonOccludedCollisionStop if there are vehicle within the margin for some
+   * duration from ego's entry to yield
+   * @attention this function has access to value() of
+   * intersection_stoplines.occlusion_peeking_stopline
+   */
+  std::optional<intersection::NonOccludedCollisionStop> isGreenPseudoCollisionStatus(
+    const autoware_auto_planning_msgs::msg::PathWithLaneId & path,
+    const size_t collision_stopline_idx,
+    const intersection::IntersectionStopLines & intersection_stoplines,
+    const TargetObjects & target_objects);
+
+  /**
    * @brief check collision
    */
   bool checkCollision(
     const autoware_auto_planning_msgs::msg::PathWithLaneId & path, TargetObjects * target_objects,
-    const PathLanelets & path_lanelets, const size_t closest_idx,
+    const intersection::PathLanelets & path_lanelets, const size_t closest_idx,
     const size_t last_intersection_stopline_candidate_idx, const double time_delay,
     const TrafficPrioritizedLevel & traffic_prioritized_level);
 
@@ -750,7 +708,6 @@ private:
   void cutPredictPathWithDuration(TargetObjects * target_objects, const double time_thr);
 
   /**
-   * @fn
    * @brief calculate ego vehicle profile along the path inside the intersection as the sequence of
    * (time of arrival, traveled distance) from current ego position
    */
@@ -758,28 +715,7 @@ private:
     const autoware_auto_planning_msgs::msg::PathWithLaneId & path, const size_t closest_idx,
     const size_t last_intersection_stopline_candidate_idx, const double time_delay,
     tier4_debug_msgs::msg::Float64MultiArrayStamped * debug_ttc_array);
-
-  std::vector<lanelet::ConstLineString3d> generateDetectionLaneDivisions(
-    lanelet::ConstLanelets detection_lanelets,
-    const lanelet::routing::RoutingGraphPtr routing_graph_ptr, const double resolution);
-
-  /**
-   * @fn
-   * @brief check occlusion status
-   */
-  OcclusionType getOcclusionStatus(
-    const std::vector<lanelet::CompoundPolygon3d> & attention_areas,
-    const lanelet::ConstLanelets & adjacent_lanelets,
-    const lanelet::CompoundPolygon3d & first_attention_area,
-    const util::InterpolatedPathInfo & interpolated_path_info,
-    const std::vector<lanelet::ConstLineString3d> & lane_divisions,
-    const TargetObjects & target_objects);
-
-  /*
-   * @fn
-   * @brief check if associated traffic light is green
-   */
-  bool isGreenSolidOn(lanelet::ConstLanelet lane);
+  /** @} */
 
   /*
   bool IntersectionModule::checkFrontVehicleDeceleration(
@@ -789,7 +725,7 @@ private:
     const double assumed_front_car_decel);
   */
 
-  DebugData debug_data_;
+  mutable DebugData debug_data_;
   rclcpp::Publisher<std_msgs::msg::String>::SharedPtr decision_state_pub_;
   rclcpp::Publisher<tier4_debug_msgs::msg::Float64MultiArrayStamped>::SharedPtr ego_ttc_pub_;
   rclcpp::Publisher<tier4_debug_msgs::msg::Float64MultiArrayStamped>::SharedPtr object_ttc_pub_;
